requirements.txt
discord.py
ipdb
python-dotenv
discord-py-interactions
